+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-express/index.ts
@@ -104,6 +104,25 @@ export default Shopware.Component.wrapComponentConfig({
 
             return criteria;
         },
+
+        ecsShippingCallbackSettingDisabled(): boolean {
+            return this.renderSettingsDisabled || !!this.actualConfigData['SwagPayPal.settings.isLocalEnvironment'];
+        },
+
+        ecsShippingCallbackSettingTooltip(): { message: string; showOnDisabledElements?: boolean; disabled?: boolean } {
+            if (this.actualConfigData['SwagPayPal.settings.isLocalEnvironment']) {
+                return {
+                    message: this.$t('swag-paypal-settings.express.ecsShippingCallbackSettingDisabledTooltip'),
+                    showOnDisabledElements: true,
+                    disabled: false,
+                };
+            }
+
+            return {
+                message: '',
+                disabled: true,
+            };
+        },
     },
 
     watch: {